+++ drink-cli/src/app_state/print.rs
@@ -1,4 +1,4 @@
-use drink::contract_api::decode_debug_buffer;
+use drink::contracts_api::decode_debug_buffer;
 use pallet_contracts::ContractResult;
 use ratatui::{
     style::{Color, Modifier, Style},
