+++ drink/src/runtime.rs
@@ -2,9 +2,9 @@
 //! `drink` with any runtime that implements the `Runtime` trait.
 
 pub mod minimal;
+pub mod pallet_contracts_debugging;
 pub use frame_metadata::RuntimeMetadataPrefixed;
 pub use minimal::MinimalRuntime;
-pub mod pallet_contracts_debugging;
 
 /// The type of an account identifier.
 pub type AccountIdFor<R> = <R as frame_system::Config>::AccountId;
