+++ drink/src/sandbox.rs
@@ -3,7 +3,7 @@
 mod sandbox_config;
 pub use sandbox_config::SandboxConfig;
 pub mod balance_api;
-pub mod contract_api;
+pub mod contracts_api;
 pub mod runtime_api;
 pub mod system_api;
 pub mod timestamp_api;
