+++ drink/src/sandbox/contracts_api.rs
@@ -2,19 +2,22 @@
 use std::ops::Not;
 
 use frame_support::{traits::fungible::Inspect, weights::Weight};
-use frame_system::Config;
+use frame_system::Config as SysConfig;
 use pallet_contracts::{
     Code, CodeUploadResult, CollectEvents, ContractExecResult, ContractInstantiateResult,
     DebugInfo, Determinism,
 };
 use parity_scale_codec::Decode as _;
 
-use crate::{runtime::AccountIdFor, EventRecordOf, Sandbox, SandboxApi};
+use crate::{runtime::AccountIdFor, EventRecordOf, Sandbox};
 
 type BalanceOf<R> =
     <<R as pallet_contracts::Config>::Currency as Inspect<AccountIdFor<R>>>::Balance;
 
-pub trait ContractsApi<R: pallet_contracts::Config>: SandboxApi<R> {
+impl<Config: crate::SandboxConfig> Sandbox<Config>
+where
+    Config::Runtime: pallet_contracts::Config,
+{
     /// Interface for `bare_instantiate` contract call with a simultaneous upload.
     ///
     /// # Arguments
@@ -26,19 +29,23 @@ pub trait ContractsApi<R: pallet_contracts::Config>: SandboxApi<R> {
     /// * `origin` - The sender of the contract call.
     /// * `gas_limit` - The gas limit for the contract call.
     /// * `storage_deposit_limit` - The storage deposit limit for the contract call.
-    #[allow(clippy::too_many_arguments)]
-    fn deploy_contract(
+    #[allow(clippy::type_complexity, clippy::too_many_arguments)]
+    pub fn deploy_contract(
         &mut self,
         contract_bytes: Vec<u8>,
-        value: BalanceOf<R>,
+        value: BalanceOf<Config::Runtime>,
         data: Vec<u8>,
         salt: Vec<u8>,
-        origin: AccountIdFor<R>,
+        origin: AccountIdFor<Config::Runtime>,
         gas_limit: Weight,
-        storage_deposit_limit: Option<BalanceOf<R>>,
-    ) -> ContractInstantiateResult<AccountIdFor<R>, BalanceOf<R>, EventRecordOf<R>> {
-        self.externalities().execute_with(|| {
-            pallet_contracts::Pallet::<R>::bare_instantiate(
+        storage_deposit_limit: Option<BalanceOf<Config::Runtime>>,
+    ) -> ContractInstantiateResult<
+        AccountIdFor<Config::Runtime>,
+        BalanceOf<Config::Runtime>,
+        EventRecordOf<Config::Runtime>,
+    > {
+        self.externalities.execute_with(|| {
+            pallet_contracts::Pallet::<Config::Runtime>::bare_instantiate(
                 origin,
                 value,
                 gas_limit,
@@ -63,26 +70,31 @@ pub trait ContractsApi<R: pallet_contracts::Config>: SandboxApi<R> {
     /// * `origin` - The sender of the contract call.
     /// * `gas_limit` - The gas limit for the contract call.
     /// * `storage_deposit_limit` - The storage deposit limit for the contract call.
-    #[allow(clippy::too_many_arguments)]
-    fn instantiate_contract(
+    #[allow(clippy::type_complexity, clippy::too_many_arguments)]
+    pub fn instantiate_contract(
         &mut self,
         code_hash: Vec<u8>,
-        value: BalanceOf<R>,
+        value: BalanceOf<Config::Runtime>,
         data: Vec<u8>,
         salt: Vec<u8>,
-        origin: AccountIdFor<R>,
+        origin: AccountIdFor<Config::Runtime>,
         gas_limit: Weight,
-        storage_deposit_limit: Option<BalanceOf<R>>,
-    ) -> ContractInstantiateResult<AccountIdFor<R>, BalanceOf<R>, EventRecordOf<R>> {
+        storage_deposit_limit: Option<BalanceOf<Config::Runtime>>,
+    ) -> ContractInstantiateResult<
+        AccountIdFor<Config::Runtime>,
+        BalanceOf<Config::Runtime>,
+        EventRecordOf<Config::Runtime>,
+    > {
         let mut code_hash = &code_hash[..];
-        self.externalities().execute_with(|| {
-            pallet_contracts::Pallet::<R>::bare_instantiate(
+        self.externalities.execute_with(|| {
+            pallet_contracts::Pallet::<Config::Runtime>::bare_instantiate(
                 origin,
                 value,
                 gas_limit,
                 storage_deposit_limit,
                 Code::Existing(
-                    <R as Config>::Hash::decode(&mut code_hash).expect("Invalid code hash"),
+                    <Config::Runtime as SysConfig>::Hash::decode(&mut code_hash)
+                        .expect("Invalid code hash"),
                 ),
                 data,
                 salt,
@@ -99,15 +111,16 @@ pub trait ContractsApi<R: pallet_contracts::Config>: SandboxApi<R> {
     /// * `contract_bytes` - The contract code.
     /// * `origin` - The sender of the contract call.
     /// * `storage_deposit_limit` - The storage deposit limit for the contract call.
-    fn upload_contract(
+    pub fn upload_contract(
         &mut self,
         contract_bytes: Vec<u8>,
-        origin: AccountIdFor<R>,
-        storage_deposit_limit: Option<BalanceOf<R>>,
+        origin: AccountIdFor<Config::Runtime>,
+        storage_deposit_limit: Option<BalanceOf<Config::Runtime>>,
         determinism: Determinism,
-    ) -> CodeUploadResult<<R as frame_system::Config>::Hash, BalanceOf<R>> {
-        self.externalities().execute_with(|| {
-            pallet_contracts::Pallet::<R>::bare_upload_code(
+    ) -> CodeUploadResult<<Config::Runtime as frame_system::Config>::Hash, BalanceOf<Config::Runtime>>
+    {
+        self.externalities.execute_with(|| {
+            pallet_contracts::Pallet::<Config::Runtime>::bare_upload_code(
                 origin,
                 contract_bytes,
                 storage_deposit_limit,
@@ -127,18 +140,18 @@ pub trait ContractsApi<R: pallet_contracts::Config>: SandboxApi<R> {
     /// * `gas_limit` - The gas limit for the contract call.
     /// * `storage_deposit_limit` - The storage deposit limit for the contract call.
     #[allow(clippy::too_many_arguments)]
-    fn call_contract(
+    pub fn call_contract(
         &mut self,
-        address: AccountIdFor<R>,
-        value: BalanceOf<R>,
+        address: AccountIdFor<Config::Runtime>,
+        value: BalanceOf<Config::Runtime>,
         data: Vec<u8>,
-        origin: AccountIdFor<R>,
+        origin: AccountIdFor<Config::Runtime>,
         gas_limit: Weight,
-        storage_deposit_limit: Option<BalanceOf<R>>,
+        storage_deposit_limit: Option<BalanceOf<Config::Runtime>>,
         determinism: Determinism,
-    ) -> ContractExecResult<BalanceOf<R>, EventRecordOf<R>> {
-        self.externalities().execute_with(|| {
-            pallet_contracts::Pallet::<R>::bare_call(
+    ) -> ContractExecResult<BalanceOf<Config::Runtime>, EventRecordOf<Config::Runtime>> {
+        self.externalities.execute_with(|| {
+            pallet_contracts::Pallet::<Config::Runtime>::bare_call(
                 origin,
                 address,
                 value,
@@ -169,7 +182,7 @@ mod tests {
 
     use super::*;
     use crate::{
-        minimal::RuntimeEvent, runtime::Runtime, session::NO_SALT, MinimalRuntime,
+        minimal::RuntimeEvent, sandbox::SandboxConfig, session::NO_SALT, MinimalRuntime,
         DEFAULT_GAS_LIMIT,
     };
